100l to whoever wrote this crap using 1/10000 units. it caused framerates to get trashed from 30000/1001 to 2997/100, etc.!
git-svn-id: svn://svn.mplayerhq.hu/mplayer/trunk@16185 b3059339-0415-0410-9bf9-f77b7e298cf2

+++ libmpdemux/mpeg_hdr.h
@@ -6,7 +6,7 @@ typedef struct {
     int display_picture_height;
     int aspect_ratio_information;
     int frame_rate_code;
-    int fps; // fps*10000
+    float fps;
     int bitrate; // 0x3FFFF==VBR
     // timing:
     int picture_structure;
